change decimal and add API:setLogoutCallback

+++ Assets/MirrorMarketSDK/Runtime/Models/Request/Marketplace/CancelNftListOnMarketplaceRequest.cs
@@ -9,7 +9,7 @@ namespace MirrorworldSDK.Models
     {
         public string mint_address;
 
-        public decimal price;
+        public float price;
 
         public string confirmation;
     }
