+++ Assets/MirrorMarketSDK/Runtime/Models/Request/Marketplace/ListNftOnMarketplaceRequest.cs
@@ -8,6 +8,6 @@ namespace MirrorworldSDK.Models
     {
         public string mint_address;
 
-        public decimal price;
+        public float price;
     }
 }
